pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "macs-pipeline"
version = "0.1.0"
description = "Modular implementation of MACS from 'Attention Consistency for LLMs Explanation'."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "OpenAI Codex" }
]
dependencies = [
  "matplotlib>=3.8",
  "numpy>=1.26",
  "pillow>=10.0",
  "scipy>=1.11",
  "torch>=2.4",
  "transformers>=4.47"
]

[project.optional-dependencies]
vl = [
  "qwen-vl-utils>=0.0.8"
]

[project.scripts]
macs = "macs.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
